addressing CR
