+++ crypto/ml_kem/ml_kem_ipd_ref_common/cbd.c
@@ -1,4 +1,6 @@
 #include <stdint.h>
+#include <assert.h>
+
 #include "params.h"
 #include "cbd.h"
