+++ crypto/ml_kem/ml_kem_ipd_ref_common/params.h
@@ -2,7 +2,6 @@
 #define PARAMS_H
 
 #include <openssl/base.h>
-#include <assert.h>
 
 // The only defined parameters are those that don't depend
 // on the parameter set. All other parameters are specified
@@ -35,43 +34,6 @@ typedef struct {
   size_t ciphertext_bytes;
 } ml_kem_params;
 
-static void ml_kem_params_init(ml_kem_params *params, size_t k) {
-  assert((k == 2) || (k == 3) || (k == 4));
-
-  size_t eta1 = (k == 2) ? 3 : 2;
-  size_t poly_compressed_bytes = (k == 4) ? 160 : 128;
-  size_t poly_vec_bytes = k * KYBER_POLYBYTES;
-  size_t poly_vec_compressed_bytes = (k == 4) ? (352 * k) : (320 * k);
-  size_t indcpa_public_key_bytes = poly_vec_bytes + KYBER_SYMBYTES;
-  size_t indcpa_secret_key_bytes = poly_vec_bytes;
-  size_t indcpa_bytes = poly_vec_compressed_bytes + poly_compressed_bytes;
-  size_t public_key_bytes = indcpa_public_key_bytes;
-  size_t secret_key_bytes = indcpa_secret_key_bytes + indcpa_public_key_bytes + 2*KYBER_SYMBYTES;
-  size_t ciphertext_bytes = indcpa_bytes;
-
-  params->k = k;
-  params->eta1 = eta1;
-  params->poly_compressed_bytes = poly_compressed_bytes;
-  params->poly_vec_bytes = poly_vec_bytes;
-  params->poly_vec_compressed_bytes = poly_vec_compressed_bytes;
-  params->indcpa_public_key_bytes = indcpa_public_key_bytes;
-  params->indcpa_secret_key_bytes = indcpa_secret_key_bytes;
-  params->indcpa_bytes = indcpa_bytes;
-  params->public_key_bytes = public_key_bytes;
-  params->secret_key_bytes = secret_key_bytes;
-  params->ciphertext_bytes = ciphertext_bytes;
-}
-
-OPENSSL_UNUSED static void ml_kem_512_params_init(ml_kem_params *params) {
-  ml_kem_params_init(params, 2);
-}
-OPENSSL_UNUSED static void ml_kem_768_params_init(ml_kem_params *params) {
-  ml_kem_params_init(params, 3);
-}
-OPENSSL_UNUSED static void ml_kem_1024_params_init(ml_kem_params *params) {
-  ml_kem_params_init(params, 4);
-}
-
 // We define max values for some parameters because they are used
 // for static allocation.
 #define KYBER_K_MAX (4)
@@ -83,5 +45,9 @@ OPENSSL_UNUSED static void ml_kem_1024_params_init(ml_kem_params *params) {
 #define KYBER_CIPHERTEXTBYTES_MAX (KYBER_INDCPA_BYTES_MAX)
 
 #define KYBER_NAMESPACE(s) ml_kem_##s##_ref
+
+void ml_kem_512_params_init(ml_kem_params *params);
+void ml_kem_768_params_init(ml_kem_params *params);
+void ml_kem_1024_params_init(ml_kem_params *params);
 
 #endif
